+++ Symfony/CS/Fixer/PSR2/BracesFixer.php
@@ -460,7 +460,7 @@ class BracesFixer extends AbstractFixer
             }
         }
 
-        throw new \RuntimeExpcetion('Statement end not found');
+        throw new \RuntimeException('Statement end not found');
     }
 
     private function getClassyTokens()