+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Changelog for v1.4
 ------------------
 
 * feature #841 PhpdocParamsFixer: added aligning var/type annotations (GrahamCampbell)
+* bug #948 NewWithBracesFixer - fix case with multidimensional array (keradus)
+* bug #945 Skip files containing __halt_compiler() on PHP 5.3 (stof)
+* bug #946 BracesFixer - fix typo in exception name (keradus)
+* bug #940 Tokens::setCode - apply missing transformation (keradus)
 * bug #908 BracesFixer - fix invalide inserting brace for control structure without brace and lambda inside of it (keradus)
 * bug #903 NoEmptyLinesAfterPhpdocsFixer - fix bug with Windows style lines (GrahamCampbell)
 * bug #895 [PSR-2] Preserve blank line after control structure opening brace (marcaube)